Suppress PossiblyBadArrayAssignment warning
Because of the possibility of an ArrayStoreException, JML always warns us when placing a BigInteger inside a BigInteger[].
This happens even though the BigInteger[] is defined as such and BigInteger doesn't have any subclasses.

This doesn't happen with ints, or Integers, which are final.
That could be the source of this issue.

However, in slides/demos/java/badarrayassignment, the warning could be suppressed using

//@ requires \typeof(<array>) == \type(BigInteger[]);

In slides/demos/java/rpncalculator, this didn't work.
Instead, this rather hacky solution "fixed" the problem:

//@ requires \typeof(<array>) == \typeof(new BigInteger[1]);

I'm not sure if this is a bug in OpenJML or if I'm just missing the forest for the trees.